Update README.md
Create a 3D illustration of an
animated character sitting casually
on top of a social media logo
"SocialMedia". The character must
wear casual modern clothing such
as jeans jacket and sneakers shoes.
The background of the image is a
social media profile page with a
user name "YourName" and a profile
picture that match.